Machine Learning Community(Moderated)
In order to achive the best possible results, machine learning projects often involve the testing of multiple algorithms. When it comes to experimenting with multiple algorithms, it can be quite time-consuming. When usin Python's lazy predict library,many of the most populer algorithms that are commonly used in data science can be tested in a single step!